# Conflicts:
#	src/main/resources/templates/404.html
#	src/main/resources/templates/blank.html
#	src/main/resources/templates/charts.html
#	src/main/resources/templates/forgot-password.html
#	src/main/resources/templates/index.html
#	src/main/resources/templates/login.html
#	src/main/resources/templates/register.html
#	src/main/resources/templates/tables.html

+++ src/main/java/com/woniu/yoga/communicate/dao/CommentMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 @Repository
 public interface CommentMapper {
-    int deleteByPrimaryKey(Long commentId);
+    int deleteByPrimaryKey(Integer commentId);
 
     int insert(Comment record);
